HaverErrorReport {Haver}	R Documentation

HaverErrorReport Object

Description

A HaverErrorReport object is returned by haver.data or haver.metadata as soon as one series code or database cannot be found.

Usage

## S3 method for class 'HaverErrorReport'
print(x)

Arguments

x
a HaverErrorReport object.

Details

You can access various lists of series codes pertaining to HaverErrorReport objects using the function haver.codelists. This aides for example in fixing up queries that have failed.

Value

print prints information contained in the object and returns it (invisibly).

See Also

Haver_package
haver.path
haver.data
haver.metadata
HaverData
HaverMetaData
haver.codelists
haver.datamd
haver.setalias
haver.limits
Haver_advanced
Haver_legacy

Examples

haver.path()
haver.path("examples")

## The database 'HAVERD' is found but it does not contain series 'wrong1' and 'wrong2':
her <- haver.data(codes=c("wrong1", "wrong2", "fxtwb"), dat="haverd")
her

## The list of codes that caused the query to fail as well as the list of good codes
## can be accessed with the 'haver.codelists' function.
## Empty lists are returned as zero-length character vectors.
haver.codelists(her)

## If we can do without the codes that could not be found, we can query the
## data of the good codes immediately:
hd <- haver.data(codes=haver.codelists(her)[["codes.found"]])

haver.path("restore")

[Package Haver version 2.4 Index]