dlxGetInfo {Haver}	R Documentation

Get Detailed Information about a Haver Time Series (legacy function / help entry)

Description

Get detailed information about a Haver time series.

Usage

dlxGetInfo(varname, database)

Arguments

varname
A single Haver variable name.

database
The Haver database in which the series you are examining resides.

Value

Returns a list with the following fields defined.

database: The full path to the database.

varname: The time series variable name.

startDate: A Date object. For weekly, monthly, quarterly, and annual series, this will be the last day in the period.

endDate: A Date object. For weekly, monthly, quarterly, and annual series, this will be the last day in the period.

numobs: The number of observations in the series.

frequency: A frequency code. 10=Annual, 30=Quarterly, 40=Monthly, 51=Weekly Monday ... 57=Weekly Sunday, 60=Daily

dateTimeMod: A POSIXct object showing the time the series was last modified.

magnitude: Magnitude

decPrecision: The decimal precision.

difType: An integer value. 1 if data can contain negative values, 0 otherwise.

aggType: An integer value. 1=Average, 2=Sum, 3=End of period, 9=Undefined, 0=not set

dataType: One of US$, LocCur, US$/LC, LC/US$, INDEX, Units,

geography1: Primary geography code.

geography2: Secondary geography code.

descriptor: The series description.

shortSource: The abbreviated name of the time series publisher.

longSource: The full name of the time series publisher.

Examples

## Not run: 
    dlxGetInfo("gdp", "usecon")

## End(Not run)

[Package Haver version 2.4 Index]