haver.datamd {Haver}	R Documentation

Get Metadata of an Existing HaverData Object

Description

haver.datamd returns the metadata of series contained in a HaverData object.

Usage

haver.datamd(x)

Arguments

x
A HaverData object

Details

‘md’ in the function name haver.datamd stands for ‘metadata’. Read ‘datamd’ as ‘the data's metadata’.

The metadata for all time series of a HaverData object are stored as a HaverMetaData object within the HaverData object. haver.datamd returns a copy of this object.

Value

haver.datamd returns a HaverMetaData object.

See Also

Haver_package
haver.data
haver.metadata
HaverData
HaverMetaData
HaverErrorReport
haver.codelists
haver.setalias
haver.limits
Haver_advanced
Haver_legacy

Examples

haver.path()
haver.path("examples")

## We pull in time series data, and look at the associated metadata afterwards:
hd <- haver.data(codes=c("haverd:fxtwb", "haverw:lic")
                 , freq="quart", aggmode="relaxed")
hmd <- haver.datamd(hd)
hmd

print(hmd, fields=NA)

haver.path("restore")

[Package Haver version 2.4 Index]