haver.metadata {Haver}	R Documentation

Access Haver Analytics Time Series Metadata

Description

haver.metadata queries metadata information on time series contained in Haver Analytics databases. The information is returned either as a HaverMetaData object or as a data.frame.

This help entry assumes that you are familiar with Haver_package.

Usage

haver.metadata(codes, database, decode = TRUE, cbreak = 500)

Arguments

codes
is either a character vector or it is an object that fulfills the conditions of
haver.hasfullcodes. If a character vector is supplied, the elements can be regular series codes or additionally supply the database where the code resides, in which case the elements must be specified in database:seriescode format. Examples of character vector elements are ‘gdp’ and ‘usecon:gdp’. Haver series codes may also be supplied through an object that passes haver.hasfullcodes. The only difference of usage of the codes argument in comparison to haver.data is that it IS optional here (see section ‘Details’ below).

database
is a character string. This argument is optional and specifies a default Haver database (excluding the path to the database, and excluding any file extension). If this argument is omitted, all elements of ‘codes’ must be in database:seriescode format. Otherwise the function errors out.

decode
is a logical TRUE / FALSE. This option is rarely used. By default (decode=TRUE), metadata information that is numerically encoded will be translated to meaningful character strings. For example, internally the aggregation type ‘AVG’ (average) is encoded as ‘1’, and the default decode=TRUE will re-translate this into ‘AVG’.

cbreak
is an integer. This option is rarely needed and specifies the number of codes after which each C-level function of the Haver package checks for user interrupts. If you want to switch off checking for interrupts, specify a very large value, such as cbreak=1e8.

Details

Argument codes may be omitted if the argument database is supplied, in which case the request is interpreted as ‘all series from database database’. Note that this is not possible with haver.data.

There are three possible combinations of using arguments codes and database.

only supply codes. In this case all elements of codes have to be in database:seriescode format.

only supply database. This will retrieve metadata for all series contained in database.

supply both codes and database. Metadata for all elements of codes is queried. If an element of codes is not in database:seriescode format, the value of database will serve as the default database.

The order of series records within the returned object corresponds to the order of series in the input argument codes.

As is generally the case with Haver functions, (character) values of arguments supplied are not case-sensitive.

Value

An object of class HaverMetaData. They contain various fields with information on series metadata. For a detailed description of these fields, see HaverMetaData.

In the case of invalid series code and/or database specifications, a HaverErrorReport is returned.

Query Errors

As soon as one series code or one database specified in a query cannot be found, haver.metadata returns a HaverErrorReport. This object contains details on what went wrong with the query. The function haver.codelists enables you to get at lists of series codes that could not be found.

See Also

Haver_package
haver.path
haver.data
HaverData
HaverMetaData
HaverErrorReport
haver.codelists
haver.datamd
haver.setalias
haver.limits
Haver_advanced
Haver_legacy

Examples

haver.path()
haver.path("examples")

## The following statements illustrate how you can specify series codes and
## database names. Argument usage of 'codes' and 'database' is identical to
## the statements of the 'Examples' section in 'haver.data'.
## All queries below are equivalent.
haver.metadata(cod=c(       "fxtwb",        "fxtwm",        "fxtwotp"), dat="haverd"  )
haver.metadata(cod=c(       "FXTWB",        "FXTWM",        "FXTWOTP"), dat="HAVERD"  )
haver.metadata(cod=c("haverd:fxtwb",        "fxtwm",        "fxtwotp"), dat="haverd"  )
haver.metadata(cod=c("haverd:fxtwb", "haverd:fxtwm", "haverd:fxtwotp")                )
haver.metadata(cod=c("haverd:fxtwb", "haverd:fxtwm", "haverd:fxtwotp"), dat="havermqa")

## Querying codes from multiple databases is possible:
hmd_1 <- haver.metadata( cod=c("haverD:fxtwb", "haverW:lic", "fcm1") , dat="haverMQA")
hmd_1

## Querying all codes of a database is not possible with 'haver.data',
## but it is with 'haver.metadata':
hmd_2 <- haver.metadata( dat="haverd" )
hmd_2

## Note that when querying an entire database, you must name the
## argument 'database' explicitly.
## Not run: 
haver.metadata("haverd")

## End(Not run)

## The statement above produces an error since the first argument of
## 'haver.metadata' is 'codes', but we meant to specify 'database',
## so to accomplish what we want to do we have to run:
haver.metadata(database="haverd")

## When setting the argument 'decode=FALSE', some metadata remain encoded as
## numeric values. In rare cases it may be helpful to look at these numeric
## values. For example:
hmd_w <- haver.metadata(database="haverw", decode=FALSE)
table(hmd_w[["frequency"]])

## shows a frequency count of the series in 'HAVERW'. 9 of those series have
## a release day-of-week of 53, 12 series of 56. These values correspond to
## Wednesday and Saturday, respectively (see 'HaverMetaData'). This can
## be confirmed by looking at one of the fields 'startdate' or 'enddate':
table(weekdays(hmd_w[["enddate"]]))

haver.path("restore")

[Package Haver version 2.4 Index]