haver.path {Haver}	R Documentation

Set and Query the Path to Haver Databases

Description

haver.path sets and queries the path to Haver databases. A correct path setting is a prerequisite for data and metadata queries run by haver.data and haver.metadata

Usage

haver.path()
haver.path(set)
haver.path("auto")
haver.path("examples")
haver.path("restore")

Arguments

set
a character string specifying the directory where Haver databases reside. ‘set’ can also be set to one of the strings ‘auto’, ‘examples’, and ‘restore’, in which case the database path is automatically set to an appropriate directory.

Details

If haver.path is called without arguments, the current setting is returned.

If set="auto", haver.path tries to set the proper path to Haver databases automatically. In some instances this may not work, depending on the R configuration and the DLX configuration of your machine.

Alternatively, you can set the database path manually. In this case you should supply the full (absolute) path of the directory where the Haver databases reside as the value for the set argument. If you supply a relative file path it is taken to be relative to the working directory. Since this is unlikely to be what you meant, a warning is issued.

haver.path("examples") will set the database directory to the directory where the example databases of the Haver package are located. Before doing so, haver.path will store the existing setting. You can re-enable this setting by issuing haver.path("restore").

Value

When argument set is not supplied, a character string that holds the current setting of the database path, and NULL otherwise.

R Startup, Loading of Package

When Haver gets attached to the R search path, it first tries to restore the database path setting from the last session. If this is not successful, it tries to detect the correct setting from the DLX setup of your machine. If neither of these attempts are successful, you have to set the database path manually.

Setting the Database Path Manually

Note that when manually setting the database path, you must supply directory separators using ‘\\’ or ‘/’. The function will fail if directories are separated by a single ‘\’.

If the supplied database path does not exist, an error occurs. If the supplied database path does not contain any Haver Analytics databases, a warning is issued.

Running ‘Examples’ Sections of Help Entries of the Haver Package

At the top of the examples section in each help entry the database path is set to point to the example databases:
> haver.path("examples").
At the end of each examples section, the previous setting is restored:
> haver.path("restore").

If you re-set the Haver database path manually after issuing haver.path("examples"), a call to haver.path("restore") will not change this setting anymore.

If you quit R or unload Haver before restoring the setting to the actual Haver Analytics databases, Haver will still set the path to the actual Haver Analytics databases when R is invoked (i.e. when Haver is attached).

See Also

Haver_package
haver.data
haver.metadata
HaverData
HaverMetaData
HaverErrorReport
haver.codelists
haver.datamd
haver.setalias
haver.limits
Haver_advanced
Haver_legacy

Examples

## query the current setting
ex_origpath <- haver.path()

## let 'haver.path' try to determine the correct setting automatically
haver.path("auto")

## set to examples databases
haver.path("examples")

## restore previous setting
haver.path("restore")

## manually restore to setting before running examples section
if (!is.null(ex_origpath)) haver.path(ex_origpath)

## Not run: 
## manually set the database path to an explicitly given directory
haver.path("c:/dlx/data")

## End(Not run)

[Package Haver version 2.4 Index]